sys/v4l2/gstv4l2src.c: Don't use bytesused/length of a buffer for GST_BUFFER_SIZE(), since length is the allocated le...
Original commit message from CVS:
* sys/v4l2/gstv4l2src.c: (gst_v4l2src_get):
Don't use bytesused/length of a buffer for GST_BUFFER_SIZE(),
since length is the allocated length in the kernel, which may
be more, and bytesused is not set by all drivers and may be
less if part of the frame was lost. Instead, use the calculated
piximage size. Fixes #301718.